レビュー指摘No.2対応
