+++ src/pages/TaskManager.tsx
@@ -1,15 +1,15 @@
-import TaskList from '../components/TaskList.tsx'
-import TaskForm from '../components/TaskForm.tsx'
 import { Box } from '@mui/material'
+import TaskFormProvider from '../components/TaskFormProvider.tsx'
+import TaskListProvider from '../components/TaskListProvider.tsx'
 
 function TaskManager() {
     return (
         <>
-            <Box sx={{ alignItems: 'flex-start', marginBottom: 3 }}>
-                <TaskForm mode={0} />
+            <Box sx={{ marginBottom: 3 }}>
+                <TaskFormProvider mode={0} />
             </Box>
             <Box>
-                <TaskList />
+                <TaskListProvider />
             </Box>
         </>
     )
